simplifies map component
