+++ assets/vue/utils/colors.ts
@@ -20,4 +20,3 @@ export const TextColors = new Map()
   .set(Colors.red, Colors.white)
   .set(Colors.purple, Colors.white)
   .set(Colors.maroon, Colors.white);
-
